revert deploy scripts
