+++ lib/sequence/setup.ts
@@ -1,8 +1,30 @@
 import { BigNumber } from "ethers";
 import { parseEther } from "ethers/lib/utils";
 
-import { expandable } from "../internal/core";
-import { Sequence, SetupContext } from "../internal/types";
+import { NeokingdomDAO, expandable } from "../internal/core";
+import {
+  ContextGenerator,
+  ContractContext,
+  Contributor,
+  NeokingdomContracts,
+  Sequence,
+} from "../internal/types";
+
+export type SetupContext = ContractContext & {
+  contributors: Contributor[];
+};
+
+export function generateSetupContext(contributors: Contributor[]) {
+  async function _generateSetupContext(n: NeokingdomDAO) {
+    const contracts = (await n.loadContractsPartial()) as NeokingdomContracts;
+    const context: SetupContext = {
+      ...contracts,
+      contributors: contributors,
+    };
+    return context;
+  }
+  return _generateSetupContext;
+}
 
 export const SETUP_SEQUENCE: Sequence<SetupContext> = [
   // Give each address one share
